+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flaw creation and update triggers a Jira task sync (OSIDB-861)
 - Config gunicorn access log file depending on environment (OSIDB-879)
 - Link tracker to flaw(s) on create/update (OSIDB-1182)
+- Implement FlawCVSS and AffectCVSS APIs with filters (OSIDB-1105)
 
 ## [3.4.2] - 2023-08-31
 ### Changed